refactor: 🔄 move Paths to a separate file
