+++ lib/get/index.ts
@@ -1,3 +1,5 @@
+import { Paths } from "../paths";
+
 type Path<T> = T extends `${infer P}.${infer Rest}` ? [P, ...Path<Rest>] : [T];
 
 type SafeObjectKeyExtractor<O, K> = K extends keyof O
@@ -31,22 +33,4 @@ type GetWithArray<O, K extends readonly any[]> = K extends readonly [infer Head,
   ? O
   : never;
 
-type Join<TKey, TPath> = TKey extends string | number
-  ? TPath extends string | number
-    ? `${TKey}${"" extends TPath ? "" : "."}${TPath}`
-    : never
-  : never;
-
-type IndicesOrder = [never, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, ...0[]];
-
-type AllPaths<TObject, TPath extends number = 10> = [TPath] extends [never]
-  ? never
-  : TObject extends object
-  ? {
-      [TKey in keyof TObject]-?: TKey extends string | number
-        ? `${TKey}` | Join<TKey, AllPaths<TObject[TKey], IndicesOrder[TPath]>>
-        : never;
-    }[keyof TObject]
-  : "";
-
-export type Get<O extends Record<any, any>, T extends AllPaths<O>> = GetWithArray<O, Path<T>>;
+export type Get<O extends Record<any, any>, T extends Paths<O>> = GetWithArray<O, Path<T>>;
